+++ packages/synapse-interface/components/Portfolio/Activity.tsx
@@ -9,9 +9,7 @@ import { Chain, Token } from '@/utils/types'
 import { tokenAddressToToken } from '@/constants/tokens'
 import { TransactionsState } from '@/slices/transactions/reducer'
 import { PortfolioState } from '@/slices/portfolio/reducer'
-import { PendingBridgeTransaction } from '@/slices/bridge/actions'
-import { BridgeState } from '@/slices/bridge/reducer'
-import { useBridgeState } from '@/slices/bridge/hooks'
+import { PendingBridgeTransaction } from '@/slices/transactions/actions'
 import { Transaction, TransactionType } from './Transaction/Transaction'
 import { PendingTransaction } from './Transaction/PendingTransaction'
 import { UserExplorerLink } from './Transaction/components/TransactionExplorerLink'
@@ -27,8 +25,8 @@ export const Activity = ({ visibility }: { visibility: boolean }) => {
     pendingAwaitingCompletionTransactions,
     fallbackQueryPendingTransactions,
     fallbackQueryHistoricalTransactions,
+    pendingBridgeTransactions,
   }: TransactionsState = useTransactionsState()
-  const { pendingBridgeTransactions }: BridgeState = useBridgeState()
   const { searchInput, searchedBalancesAndAllowances }: PortfolioState =
     usePortfolioState()
 
@@ -71,7 +69,11 @@ export const Activity = ({ visibility }: { visibility: boolean }) => {
       return true
     }
     return false
-  }, [pendingBridgeTransactions, pendingAwaitingCompletionTransactions])
+  }, [
+    pendingBridgeTransactions,
+    pendingAwaitingCompletionTransactions,
+    pendingAwaitingCompletionTransactionsWithFallback,
+  ])
 
   const hasHistoricalTransactions: boolean = useMemo(
     () => checkTransactionsExist(userHistoricalTransactions),
@@ -328,7 +330,8 @@ export const Activity = ({ visibility }: { visibility: boolean }) => {
 
 export const PendingTransactionAwaitingIndexing = () => {
   const { address } = useAccount()
-  const { pendingBridgeTransactions }: BridgeState = useBridgeState()
+  const { pendingBridgeTransactions }: TransactionsState =
+    useTransactionsState()
   return (
     <>
       {pendingBridgeTransactions.map(