+++ packages/synapse-interface/slices/portfolio/updater.tsx
@@ -1,33 +1,28 @@
 import { useEffect } from 'react'
+import { useAccount } from 'wagmi'
 import { Address } from '@wagmi/core'
-import { watchPendingTransactions, waitForTransaction } from '@wagmi/core'
 import { useAppDispatch } from '@/store/hooks'
-import { useBridgeState } from '../bridge/hooks'
 import { useTransactionsState } from '../transactions/hooks'
 import { TransactionsState } from '../transactions/reducer'
-import { BridgeState } from '../bridge/reducer'
 import { fetchAndStoreSingleNetworkPortfolioBalances } from './hooks'
-
-import { useAccount } from 'wagmi'
-import {
-  PendingBridgeTransaction,
-  updatePendingBridgeTransaction,
-} from '../bridge/actions'
+import { PendingBridgeTransaction } from '../transactions/actions'
 import { BridgeTransaction } from '../api/generated'
+import { checkTransactionsExist } from '@/utils/checkTransactionsExist'
 
 export default function Updater(): null {
   const dispatch = useAppDispatch()
   const { address } = useAccount()
-  const { pendingBridgeTransactions }: BridgeState = useBridgeState()
   const {
+    pendingBridgeTransactions,
     userHistoricalTransactions,
     isUserHistoricalTransactionsLoading,
+    fallbackQueryHistoricalTransactions,
   }: TransactionsState = useTransactionsState()
 
   // Update Origin balances when transaction resolves
   useEffect(() => {
     if (!address || !pendingBridgeTransactions) return
-    if (pendingBridgeTransactions && pendingBridgeTransactions.length > 0) {
+    if (checkTransactionsExist(pendingBridgeTransactions)) {
       const updateOriginBalancesForNewestTransaction = async () => {
         const newestTransaction: PendingBridgeTransaction =
           pendingBridgeTransactions[0]
@@ -44,10 +39,10 @@ export default function Updater(): null {
     }
   }, [pendingBridgeTransactions, address, dispatch])
 
-  // Update Destination balances for new historical transaction
+  // Update Destination balances for new historical transactions
   useEffect(() => {
     if (!address || isUserHistoricalTransactionsLoading) return
-    if (userHistoricalTransactions && userHistoricalTransactions.length > 0) {
+    if (checkTransactionsExist(userHistoricalTransactions)) {
       const updateDestinationBalancesForLastTransaction = async () => {
         const lastTransaction: BridgeTransaction = userHistoricalTransactions[0]
         const destinationChainId: number = lastTransaction.toInfo?.chainID
@@ -61,6 +56,30 @@ export default function Updater(): null {
       updateDestinationBalancesForLastTransaction()
     }
   }, [userHistoricalTransactions, address])
+
+  // Update Destination balances for new fallback historical transactions
+  useEffect(() => {
+    if (
+      !address ||
+      !checkTransactionsExist(fallbackQueryHistoricalTransactions)
+    )
+      return
+    if (checkTransactionsExist(fallbackQueryHistoricalTransactions)) {
+      const updateDestinationBalancesForNewestTransaction = async () => {
+        const newestTransaction: BridgeTransaction =
+          fallbackQueryHistoricalTransactions[0]
+        const updateChainId: number = newestTransaction?.fromInfo?.chainID
+
+        dispatch(
+          fetchAndStoreSingleNetworkPortfolioBalances({
+            address: address as Address,
+            chainId: updateChainId,
+          })
+        )
+      }
+      updateDestinationBalancesForNewestTransaction()
+    }
+  }, [fallbackQueryHistoricalTransactions])
 
   return null
 }