+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,198 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.189](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.188...@synapsecns/synapse-interface@0.1.189) (2023-11-22)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.188](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.187...@synapsecns/synapse-interface@0.1.188) (2023-11-22)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.187](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.186...@synapsecns/synapse-interface@0.1.187) (2023-11-21)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.186](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.185...@synapsecns/synapse-interface@0.1.186) (2023-11-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.185](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.184...@synapsecns/synapse-interface@0.1.185) (2023-11-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.184](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.183...@synapsecns/synapse-interface@0.1.184) (2023-11-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.183](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.182...@synapsecns/synapse-interface@0.1.183) (2023-11-20)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.182](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.181...@synapsecns/synapse-interface@0.1.182) (2023-11-15)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.181](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.180...@synapsecns/synapse-interface@0.1.181) (2023-11-11)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.180](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.179...@synapsecns/synapse-interface@0.1.180) (2023-11-10)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.179](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.178...@synapsecns/synapse-interface@0.1.179) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.178](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.177...@synapsecns/synapse-interface@0.1.178) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.177](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.176...@synapsecns/synapse-interface@0.1.177) (2023-11-09)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.176](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.175...@synapsecns/synapse-interface@0.1.176) (2023-11-08)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.175](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.174...@synapsecns/synapse-interface@0.1.175) (2023-11-08)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.174](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.173...@synapsecns/synapse-interface@0.1.174) (2023-11-08)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.173](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.172...@synapsecns/synapse-interface@0.1.173) (2023-11-07)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.172](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.171...@synapsecns/synapse-interface@0.1.172) (2023-11-07)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.171](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.170...@synapsecns/synapse-interface@0.1.171) (2023-11-07)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.170](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.169...@synapsecns/synapse-interface@0.1.170) (2023-11-07)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.169](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.168...@synapsecns/synapse-interface@0.1.169) (2023-11-07)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.168](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.167...@synapsecns/synapse-interface@0.1.168) (2023-11-06)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.167](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.166...@synapsecns/synapse-interface@0.1.167) (2023-11-06)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.166](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.165...@synapsecns/synapse-interface@0.1.166) (2023-11-06)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.165](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.164...@synapsecns/synapse-interface@0.1.165) (2023-11-04)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface